fixed selectors in actions.js did not use options
I'm using the selector style, `$(selector, parentSelector)`, which is not
my preferred style. Given that the parent selector is only available
to me as a string, I feel like this method preserves the original style
as much as possible. Internally, jQuery/sizzle just converts the second
argument to a jQuery object and runs .find() on it.